Update the routes structure to use the layout as well as the Protected routes.
